requirements.txt
Django==1.8.4
MySQL-python
Pillow
django-crispy-forms==1.5.0
python-dateutil
